Param transformer for sanitizing and normalizing
Add a param transformer that sanitizes and normalizes parameters coming
in. Calls `to_h` on ActionController::Params to get a hash of permitted
parameters and then maps their keys against a configurable list in order
to determine if any transformations should happen.

Currently the only transformed param is email, which is downcased in
order to prevent case insensitivity prevent signing in or resetting
passwords.